+++ window-tree.c
@@ -486,12 +486,12 @@ window_tree_draw_label(struct screen_write_ctx *ctx, u_int px, u_int py,
 	len = strlen(label);
 	if (sx == 0 || sy == 1 || len > sx)
 		return;
-	ox = (sx - len + 1) / 2;
+	ox = (sx - (u_int)len + 1) / 2;
 	oy = (sy + 1) / 2;
 
 	if (ox > 1 && ox + len < sx - 1 && sy >= 3) {
 		screen_write_cursormove(ctx, px + ox - 1, py + oy - 1, 0);
-		screen_write_box(ctx, len + 2, 3);
+		screen_write_box(ctx, (u_int)len + 2, 3);
 	}
 	screen_write_cursormove(ctx, px + ox, py + oy, 0);
 	screen_write_puts(ctx, gc, "%s", label);
